% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13_abund_rel_vec.R, R/25_abund_rel_vec.R
\name{abund_rel_vec}
\alias{abund_rel_vec}
\title{Calculate relative abundance (possibly depracated)}
\usage{
abund_rel_vec(traj.list, K = NA, years.vec = 10)

abund_rel_vec(traj.list, K = NA, years.vec = 10)
}
\arguments{
\item{traj.list}{a list of simulation outputs (high.list,med.list,low.lis,zero.list) each of these lists is a list of three, one output at each starting depletion (high, med, low depletion)}

\item{K}{carrying capacity}

\item{years.vec}{vector of years to check abundances at. If length = 1, the fn returns relative abundance at that year}
}
\description{
Calculate relative abundance (possibly depracated)

Calculate relative abundance (possibly depracated)
}
