% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16_add_trans.R
\name{add_trans}
\alias{add_trans}
\title{Add transparency to a color}
\usage{
add_trans(color, trans)
}
\arguments{
\item{color}{input color}

\item{trans}{degree of transparency - a numeric values between 0 and 255 where 255 is fully visible)}
}
\value{
a hex code for the transparent version of \code{color}
}
\description{
Add transparency to a color
}
\details{
This function adds transparency to a color. It is originally by Tim Essington. Define transparency with an integer between 0 and 255, 0 being fully transparent and 255 being fully visible. Works with either color and trans a vector of equal length, or one of the two of length 1.
}
\examples{
x <- add_trans("red", 100)
print(x)
}
