% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat-data.R
\docType{data}
\name{dat}
\alias{dat}
\title{Life history parameters for specific pinnipeds}
\format{
A data frame with 11 rows and 11 variables:
\describe{
  \item{Species}{Species}
  \item{Parameter}{Life history parameter}
  \item{Value}{Parameter value}
  \item{Range}{Parameter range, if a range is reported in the paper instead of a single value}
  \item{Citation}{Source for parameter value or range}
  \item{doi}{doi for source}
  \item{Location}{Location where data were collected}
  \item{Time.period}{Time period studied in the paper}
  \item{Sex.specific}{If only one sex was included in the paper}
  \item{Notes}{Messy notes about data collection and source}
  ...
}
}
\source{
Compiled from literature by M Siple and coauthors
}
\usage{
dat
}
\description{
A table with life history information for pinnipeds.
}
\keyword{datasets}
