% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_find_msyr.R
\name{find_msyr}
\alias{find_msyr}
\title{calculate MSYR}
\usage{
find_msyr(E.start, lh.params, fmax, N0)
}
\arguments{
\item{E.start}{a starting guess for bycatch mortality rate that will result in FMSY (numeric value)}

\item{lh.params}{a list of life history parameters (S0, S1plus, nages, AgeMat, lmabdaMax, K1plus, and z)}

\item{fmax}{Max theoretical fecundity (num)}

\item{N0}{unfished nums per recruit - mature adults}
}
\description{
one of the functions needed for getting MNPL. MSYR is the same as FMSY.
}
