% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_get_dc.R
\name{get_dc}
\alias{get_dc}
\title{Calculate the derivative of C in getMNPL()}
\usage{
get_dc(
  logit.E,
  S0 = S0.w,
  S1plus = S1plus.w,
  nages = nages.w,
  AgeMat = AgeMat.w,
  lambdaMax = lambdaMax.w,
  K1plus = K1plus.w,
  z = z.w
)
}
\arguments{
\item{logit.E}{logit transform of bycatch mortality}

\item{S0}{calf/pup survival}
}
\description{
this is the function to minimize, but needs to be restricted to that yield > 0
All life history params are as above
}
