% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_get_dz.R
\name{get_dz}
\alias{get_dz}
\title{Derivative of z (function to minimize)}
\usage{
get_dz(z, MNPL, lh.params)
}
\arguments{
\item{z}{the degree of compensation}

\item{MNPL}{desired maximum net productivity level}

\item{lh.params}{a list of life history parameters (juvenile survival S0, adult survival S1plus, age at maturity AgeMat, plus group age, max theoretical fecundity fmax, and maximum steady rate of increase (population growth rate) lambdaMax)}
}
\value{
The difference between the MNPL associated with the value of z that the user defined and the MNPL that the user has defined.
}
\description{
Derivative of z (function to minimize)
}
\examples{
get_dz(z = 2.39, MNPL = 0.5, 
lh.params = list(S0 = 0.944, S1plus = 0.99, AgeMat = 17, 
PlusGroupAge = 25, fmax = 0.29, lambdaMax = 1.04))
}
