% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_get_mnpl.R
\name{get_mnpl}
\alias{get_mnpl}
\title{Calculate MNPL}
\usage{
get_mnpl(E.start = 0.001, lh.params)
}
\arguments{
\item{E.start}{a starting guess for bycatch mortality rate that will result in MSYR (equivalent mathematically to FMSY for fisheries)}

\item{lh.params}{a list of life history parameters}
}
\description{
Calculate MNPL
}
