% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/19_ggradar.R
\name{ggradar}
\alias{ggradar}
\title{Radar plot}
\usage{
ggradar(
  plot.data,
  axis.labels = colnames(plot.data)[-1],
  grid.label.size = 7,
  axis.label.size = 8,
  plot.legend = if (nrow(plot.data) > 1) TRUE else FALSE,
  legend.text.size = grid.label.size,
  palette.vec = c("#D53E4F", "#FC8D59", "#FEE08B", "#E6F598", "#99D594", "#3288BD"),
  manual.levels = NA
)
}
\arguments{
\item{plot.data}{a dataframe with performance measures as columns and management scenarios as rows}

\item{axis.labels}{a vector of names for the performance measures}

\item{grid.label.size}{a numeric value for grid label size}

\item{axis.label.size}{a numeric value for axis label size}

\item{plot.legend}{logical; whether or not to plot a legend to the right of the plot}

\item{legend.text.size}{numeric value for size of legend text}

\item{palette.vec}{a vector of colors to use for the different scenarios (each row = 1 color)}

\item{manual.levels}{}
}
\description{
Radar plot
}
\details{
I modified the `ggradar()` function slightly to do nice things that I like, like use a custom color palette, including different line types, etc.

Since this code was originally written, ggradar has becomes its own standalone package. For more information and for the most current version of the function, see Ricardo Bion's \href{https://github.com/ricardo-bion/ggradar/}{GitHub}
}
\author{
Ricardo Bion, modified slightly by Margaret Siple
}
