% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_inv_logit.R
\name{inv_logit}
\alias{inv_logit}
\title{Get inverse logit of a value}
\usage{
inv_logit(x)
}
\arguments{
\item{x}{A number}
}
\value{
the inverse logit of x.
}
\description{
The inverse logit of x
}
\examples{
inv_logit(0)
inv_logit(1)
}
