% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15_plot_bycatch_guesses.R
\name{plot_bycatch_guesses}
\alias{plot_bycatch_guesses}
\title{Plot distribution of bycatch values}
\usage{
plot_bycatch_guesses(
  highval,
  medval,
  lowval,
  cv,
  set_size = 18,
  color.palette = c("forestgreen", "orange", "red"),
  lang = "en"
)
}
\arguments{
\item{highval}{The high end of the user-defined range of bycatch values}

\item{medval}{The middle of the user-defined bycatch range}

\item{lowval}{The low end of the user-defined range}

\item{cv}{CV of bycatch mortality or bycatch mortality rate}

\item{set_size}{base size to pass to \code{ggplot}}

\item{color.palette}{A vector of colors to represent bycatch levels, from low to high end of range}

\item{lang}{Language selected by the user (character, 2 letters)}
}
\value{
A \code{ggplot2} grob showing the distributions of bycatch values based on what the user entered.
}
\description{
Plot user-defined bycatch values (uses \code{geom_linerange()} from \code{ggplot2})
}
