% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/22_plot_pinnipeds.R
\name{plot_pinnipeds}
\alias{plot_pinnipeds}
\title{Plot all the pinniped life history parameters}
\usage{
plot_pinnipeds(dat, central = FALSE, set_size = 14)
}
\arguments{
\item{dat}{a dataframe with all the parameters in in (from Pinniped parameters - Data.csv)}

\item{central}{Whether or not to plot a line showing the central tendency of all the parameters}

\item{set_size}{Base size for plot}
}
\value{

}
\description{
Make the table on the documentation tab of the app, which shows the range of pinniped life history parameters
}
\examples{
plot_pinnipeds(dat = dat)
}
