% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/17_plot_proj.R
\name{plot_proj}
\alias{plot_proj}
\title{Plot projections}
\usage{
plot_proj(
  high,
  med,
  low,
  years.plot = 50,
  ylims,
  spaghetti = FALSE,
  K1plus = 9000,
  InitDepl = 0.8,
  color.palette = c("forestgreen", "darkorange", "red"),
  lang = "en"
)
}
\arguments{
\item{high}{a list containing output from \code{Projections()} (including a matrix of simulation trajectories) - corresponding to a high bycatch level (this is at the high end of the range defined by the user in the Shiny app)}

\item{med}{a list containing output from \code{Projections()} (including a matrix of simulation trajectories) - corresponding to a medium bycatch level (this is at the median of the low and high end of the range defined by the user)}

\item{low}{a list containing output from \code{Projections()} (including a matrix of simulation trajectories) - corresponding to a high bycatch level (this is at the high end of the range defined by the user)}

\item{years.plot}{number of years to plot on the x axis}

\item{ylims}{y-limits of projection plot}

\item{spaghetti}{either FALSE or a number, where the number is how many simulations to show from the same scenario}

\item{K1plus}{carrying capacity in terms of age 1+ individuals}

\item{InitDepl}{initial depletion level (set by the user)}

\item{color.palette}{a vector of three colors to use for low, medium and high bycatch rates}
}
\value{
A plot of 50 percent and 90 percent confidence intervals of population projections (if \code{spaghetti == FALSE}) or a spaghetti plot (if \code{spaghetti ==TRUE}),  from \code{Projections()}.
}
\description{
plots outputs from several projections that result from a Projections() call.
}
