% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13_pop_vs_yield.R
\name{pop_vs_yield}
\alias{pop_vs_yield}
\title{Yield or productivity curve}
\usage{
pop_vs_yield(
  z.vec = c(1, 2.39, 5.99),
  lh.params = lh.params1,
  add.legend = FALSE,
  ggp = TRUE,
  linecolor = "#123f5a",
  lang = "en"
)
}
\arguments{
\item{z.vec}{a vector of z values}

\item{lh.params}{a list of life history parameters}

\item{add.legend}{logical; whether or not to add a legend}

\item{ggp}{logical; whether to plot in ggplot and plot a single yield curve; set to \code{FALSE} for base R plot and multiple yield curves.}

\item{linecolor}{color of yield curve line}

\item{lang}{language selected by the user (character)}
}
\value{
a plot of population size vs. 'yield'
}
\description{
relative population size (x) vs. Sustainable yield (y)
}
\examples{
pop_vs_yield(lh.params = list(
  S0 = 0.944, S1plus = 0.99, AgeMat = 17,
  nages = 19, z = 2.39, lambdaMax = 1.04
), ggp = FALSE)
}
