% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/21_projections.R
\name{projections}
\alias{projections}
\title{Run multiple simulations of a marine mammal population with bycatch mortality}
\usage{
projections(
  NOut,
  ConstantBycatch = list(Catch = NA, CV = NA),
  ConstantRateBycatch = list(Rate = NA, CV = NA),
  InitDepl,
  lh.params,
  nyears,
  obs_CV = 0
)
}
\arguments{
\item{NOut}{- number of simulations}

\item{ConstantBycatch}{Mean and CV of number of animals caught as bycatch each year}

\item{ConstantRateBycatch}{Mean and CV of bycatch rate}

\item{InitDepl}{initial depletion. If obs_CV>0, this is the mean depletion}

\item{lh.params}{a list of life history parameters}

\item{nyears}{number of years to do projections}

\item{obs_CV}{observation CV. Default to 1 for simple projections}
}
\value{
list of outputs from simulations: 'trajectories' matrix has simulations
other params are the same as in the Dynamics() function
}
\description{
- generates several projections, stochasticity is in the number of catches from year to year
}
