% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_calc_z.R
\name{calc_z}
\alias{calc_z}
\title{Degree of compensation}
\usage{
calc_z(MNPL_in, lh.params_in)
}
\arguments{
\item{MNPL_in}{User-specified value for MNPL as a proportion of K (between 0 and 1)}

\item{lh.params_in}{a list of life history parameters. Must contain S0, S1plus, nages, AgeMat, lambdaMax and z.}
}
\value{
the value of z corresponding to the user-defined value of MNPL.
}
\description{
Calculate the parameter z, the degree of compensation
}
\details{
Helper function for calculating z when user specifies MNPL
}
\examples{
test.z <- calc_z(MNPL_in = 0.5,
lh.params_in = list(S0 = 0.944, S1plus = 0.99, AgeMat = 17, nages = 19,
fmax = 0.29, z = 2.39, lambdaMax = 1.04, K1plus = 9000))
test.z
}
