% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_ce.R
\name{ce}
\alias{ce}
\title{Calculate normalized sustainable yield}
\usage{
ce(
  S0 = NA,
  S1plus = NA,
  AgeMat = NA,
  nages = NA,
  z = NA,
  E = NA,
  A = NA,
  P0 = NA,
  N0 = NA
)
}
\arguments{
\item{S0}{Calf/pup survival, a numeric value between 0 and 1}

\item{S1plus}{1+ survival rate for animals age 1 year and older, a numeric value between 0 and 1}

\item{AgeMat}{Age at maturity (= age at first parturition - 1). Must be less than \code{nages}}

\item{nages}{"maximum" age, treated as the plus group age. The plus group age can be set equal to the age at maturity +2 years without losing accuracy.}

\item{z}{degree of compensation}

\item{E}{bycatch mortality rate (applies to 1+ numbers)}

\item{A}{the Pella-Tomlinson resilience parameter ((fmax - f0)/f0)}

\item{P0}{unfished number-per-recruit - 1+ adults}

\item{N0}{unfished numbers-per-recruit - mature adults}

\item{K1plus}{the pre-exploitation population size of individuals aged 1 and older. If this value is unavailable, it can be approximated by using the initial depletion and the estimate of current abundance.}
}
\value{
a single value of normalized yield for exploitation rate E
}
\description{
This function calculates the normalized sustainable yield, which is used to find MNPL (the population size at which productivity is maximized).
}
\examples{
Set parameters
S0.w = 0.5; S1plus.w = 0.944; nages.w = 25; AgeMat.w = 18 
# Get number of individuals per recruit in terms of mature individuals (N0.w)
NPROut <- npr(S0 = S0.w, S1plus = S1plus.w, nages = nages.w, AgeMat = AgeMat.w, E = 0)

N0 <- NPROut$npr # mature numbers per recruit
# Get number of individuals per recruit in terms of individuals aged 1+ (P0.w)
P0 <- NPROut$P1r # 1+ nums per recruit

ce(S0 = S0.w, S1plus = S1plus.w, 
nages = nages.w, 
AgeMat = AgeMat.w, 
E=0.01, z=2.39,A=2, N0 = N0, P0 = P0)
}
