% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20_dynamics.R
\name{dynamics}
\alias{dynamics}
\title{Generate one marine mammal population trajectory}
\usage{
dynamics(
  S0,
  S1plus,
  K1plus,
  AgeMat,
  InitDepl,
  ConstantCatch = NA,
  ConstantF = NA,
  z,
  nyears,
  nages,
  lambdaMax
)
}
\arguments{
\item{S0}{Calf/pup survival (a value between 0 and 1)}

\item{S1plus}{Survival for animals age 1 year and older (a value between 0 and 1)}

\item{K1plus}{The pre-exploitation population size of individuals aged 1 and older.  If this value is unavailable, it can be approximated by using the initial depletion and the estimate of current abundance}

\item{AgeMat}{Age at maturity in years (assumed to be age at first parturition - 1)}

\item{InitDepl}{Starting depletion level}

\item{ConstantCatch}{Total bycatch each year, expressed as a vector of length \code{nyears}}

\item{ConstantF}{vector (length = \code{nyears}) rate of bycatch each year}

\item{z}{The degree of compensation.  The default value is \code{z = 2.39}.}

\item{nyears}{Number of years to project}

\item{nages}{"Maximum" age, treated as the plus group age. The plus group age can be set equal to the age at maturity +2 years without losing accuracy}

\item{lambdaMax}{Maximum steady rate of increase (population growth rate)}
}
\value{
A list containing a matrix \code{N} of numbers at age (dimensions \code{nyears} (rows) x \code{nages} (columns)) and one vector \code{TotalPop} (a vector of length \code{nyears}), containing the number of age 1+ individuals in the population.
}
\description{
This function generates one trajectory for a marine mammal population, starting at a user-specified depletion level \code{InitDepl}.
}
\details{
The population model is a single-sex age-structured model in which the number of calves or pups born each year is density dependent, with the extent of density dependence a function of the number of mature adults \eqn{\tildeN}, the fecundity (pregnancy rate) at pre-exploitation equilibrium \eqn{f_0}, the maximum theoretical fecundity rate fmax, the degree of compensation \eqn{z}, and the abundance of individuals aged 1+ \eqn{N_{y+1}^{1+}} relative to carrying capacity \eqn{K^{1+}}. This function can be used alone but is intended to be used with \code{Projections()} to generate multiple simulations.
}
\examples{
# Generate a time series of abundance for a bowhead whale
dynamics(S0 = 0.944, S1plus = 0.99, K1plus = 9000, AgeMat = 17,
 InitDepl = 0.6, ConstantCatch = NA, ConstantF = rep(0.01, times = 100), 
 z = 2.39, nyears = 100, nages = 25, lambdaMax = 1.04)
}
