% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/28_extract_df.R
\name{extract_df}
\alias{extract_df}
\title{Extract results from `dynamics()` function as a dataframe}
\usage{
extract_df(x)
}
\arguments{
\item{x}{Outputs from a call to projections(), which are a named list}
}
\value{
a dataframe with population trajectories, initial depletion, and bycatch levels that can be printed as a table.
}
\description{
#unnest results from dynamics() and turn them into a dataframe that you can bind together
}
\examples{
x <- projections(
  NOut = 3, ConstantRateBycatch = list(Rate = 0.01, CV = 0.3),
  InitDepl = 0.8,
  lh.params = list(
    S0 = 0.944, S1plus = 0.99,
    K1plus = 9000, AgeMat = 18, nages = 20, z = 2.39, lambdaMax = 1.02
  ),
  nyears = 50, obs_CV = 0
)
extract_df(x)
}
