% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_get_diff.R
\name{get_diff}
\alias{get_diff}
\title{Calculate the derivative of C in getMNPL()}
\usage{
get_diff(
  logit.E,
  S0 = S0.w,
  S1plus = S1plus.w,
  nages = nages.w,
  AgeMat = AgeMat.w,
  lambdaMax = lambdaMax.w,
  K1plus = K1plus.w,
  z = z.w
)
}
\arguments{
\item{logit.E}{logit transform of bycatch mortality}

\item{S0}{Calf/pup survival, a numeric value between 0 and 1}

\item{S1plus}{adult survival}

\item{nages}{Plus group age in years}

\item{AgeMat}{Age at maturity in years (must be equal to or less than nages)}

\item{lambdaMax}{Maximum theoretical population growth rate}

\item{K1plus}{Carrying capacity in terms of the age 1+ component of the population}

\item{z}{Pella-Tomlinson parameter (compensation)}
}
\description{
this is the function to minimize, but needs to be restricted to that yield > 0
All life history params are as above
}
