% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_get_f.R
\name{get_f}
\alias{get_f}
\title{Get bycatch mortality rate given depletion}
\usage{
get_f(
  f.start = NA,
  S0.w = NA,
  S1plus.w = NA,
  nages.w = NA,
  AgeMat.w = NA,
  InitDepl.w = NA,
  z.w = NA,
  lambdaMax.w = NA,
  N0.w = NA,
  P0.w = NA,
  Check = F
)
}
\arguments{
\item{f.start}{an initial guess for the bycatch mortality rate E. The default value is E = 0.5}

\item{S0.w}{Calf/pup survival, a numeric value between 0 and 1. (Note: the 'w' suffix indicates that \eqn{z} is in the wrapper function, and is used inside the function by \code{optim})}

\item{S1plus.w}{survival rate for animals age 1 year and older, a numeric value between 0 and 1}

\item{nages.w}{"maximum" age, treated as the plus group age. The plus group age can be set equal to the age at maturity +2 years without losing accuracy.}

\item{AgeMat.w}{Age at maturity in years (assumed to be age at first parturition - 1)}

\item{InitDepl.w}{The depletion level to solve for as a proportion of carrying capacity; a numeric value between 0 and 1. This is equivalent to 'starting depletion'.}

\item{z.w}{The degree of compensation. The default value is \code{z = 2.39}.}

\item{lambdaMax.w}{The maximum intrinsic growth rate}

\item{N0.w}{unfished numbers per recruit in terms of mature individuals individuals}

\item{P0.w}{Number of individuals per recruit in terms of individuals aged 1+}

\item{Check}{logical; if \code{TRUE}, prints the value to be minimized (should be zero)}
}
\value{
The bycatch rate that would lead to a depletion level of \code{InitDepl.w} -- a single value.
}
\description{
This function solves for the bycatch mortality rate \eqn{E} that gives a pre-specified depletion level \code{InitDepl.w}. It is used within the \code{projections()} function to calculate the stable age distribution at which to start the projections.
}
\examples{
# Set parameters
S0.w = 0.5; S1plus.w = 0.944; nages.w = 25; AgeMat.w = 18 
InitDepl.w = 0.9; z.w = 2.39; lambdaMax.w = 1.04
# Get number of individuals per recruit in terms of mature individuals (\eqn{N0.w})
NPROut <- npr(S0 = S0.w, S1plus = S1plus.w, nages = nages.w, AgeMat = AgeMat.w, E = 0)
N0 <- NPROut$npr # mature numbers per recruit
# Get number of individuals per recruit in terms of individuals aged 1+ (\eqn{P0.w})
P0 <- NPROut$P1r # 1+ nums per recruit

# Get bycatch mortality rate for the initial depletion defined above
get_f(f.start = 0.5, 
S0.w = S0.w, S1plus.w = S1plus.w, nages.w = nages.w, AgeMat.w = AgeMat.w,
InitDepl.w = InitDepl.w, z.w = z.w, lambdaMax.w = lambdaMax.w, 
N0.w = N0, P0.w = P0,Check=T)
}
