% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_get_mnpl.R
\name{get_mnpl}
\alias{get_mnpl}
\title{Calculate MNPL}
\usage{
get_mnpl(E.start = 0.001, lh.params)
}
\arguments{
\item{E.start}{a starting guess for bycatch mortality rate that will result in MSYR (equivalent mathematically to FMSY for fisheries)}

\item{lh.params}{a list of life history parameters}
}
\description{
Calculate MNPL
}
\examples{
get_mnpl(E.start = 0.001, 
lh.params = list(S0 = 0.944, S1plus = 0.99, AgeMat = 17, nages = 19,
PlusGroupAge = 19,  fmax = 0.29, z = 2.39, lambdaMax = 1.04, 
K1plus = 9000))

}
