% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_get_rf.R
\name{get_rf}
\alias{get_rf}
\title{Get relative recruitment at bycatch mortality rate E}
\usage{
get_rf(E_in, S0, S1plus, nages, AgeMat, z, A, P0, N0)
}
\arguments{
\item{E_in}{Bycatch mortality rate}

\item{S0}{calf or pup survival (numeric value)}

\item{S1plus}{adult survival  (numeric value)}

\item{nages}{number of age classes-- including plus group age-- in years}

\item{AgeMat}{age at maturity in years}

\item{z}{degree of compensation, in the app calculated from the value of MNPL defined by the user}

\item{A}{Pella-Tomlinson resilience parameter (see Punt 1999; Annex R).
A = (FecMax - Fec0) / Fec0 (num)}

\item{P0}{unfished 1+ numbers per recruit, \eqn{tildeP(0)}}

\item{N0}{unfished mature numbers per recruit, \eqn{tildeN(0)}}
}
\value{
recruitment given exploitation rate \emph{E} - this value is multiplied by the initial abundance \eqn{N_{init}} to get initial nums at age ( a vector)
}
\description{
calculates recruitment at bycatch mortality rate \emph{E}, relative to a no-bycatch scenario
}
\examples{
S0 = 0.944; S1plus = 0.99; nages = 25; AgeMat = 17; z = 2.39; lambdaMax = 1.04;
NPROut <- npr(S0 = S0, S1plus = S1plus, nages = nages, AgeMat = AgeMat, E = 0)
N0 <- NPROut$npr # mature numbers per recruit
P0 <- NPROut$P1r # 1+ nums per recruit
Fec0 <- 1.0 / N0
FecMax <- getfecmax(S1plus = S1plus, S0 = S0, AgeMat = AgeMat, lambdaMax = lambdaMax)
A <- (FecMax - Fec0) / Fec0
get_rf(E_in = 0.01, S0 = S0, S1plus = S1plus, nages = nages, AgeMat = AgeMat, z = z, A = A,P0 = P0,N0 = N0)

}
