% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/18_multiplot_proj.R
\name{multiplot_proj}
\alias{multiplot_proj}
\title{multiplot.proj}
\usage{
multiplot_proj(
  high.d1,
  med.d1,
  low.d1,
  high.d2,
  med.d2,
  low.d2,
  high.d3,
  med.d3,
  low.d3,
  spaghetti = FALSE,
  years.to.plot = 50,
  color.palette = c("forestgreen", "darkorange", "red"),
  lang = "en"
)
}
\arguments{
\item{high.d1}{a list containing output from \code{projections()} (including a matrix of simulation trajectories) - this corresponds to the highest bycatch level (thus "high") and the depletion level ( \code{d1} indicates the lowest starting depletion level)}

\item{med.d1}{a list containing the middle bycatch value and lowest starting depletion}

\item{low.d1}{a list containing the lowest bycatch value and lowest starting depletion}

\item{high.d2}{a list containing the highest bycatch value and middle starting depletion}

\item{med.d2}{a list containing the middle bycatch value and middle starting depletion}

\item{low.d2}{a list containing the lowest bycatch value and middle starting depletion}

\item{high.d3}{a list containing the highest bycatch value and highest starting depletion}

\item{med.d3}{a list containing the middle bycatch value and highest starting depletion}

\item{low.d3}{a list containing the lowest bycatch value and highest starting depletion}

\item{spaghetti}{either FALSE or a number, where the number is how many simulations to show from the same scenario}

\item{years.to.plot}{How many years to plot on the x axis}

\item{color.palette}{A vector of length 3 giving the color values for low, medium, and high bycatch mortality or bycatch mortality rate}

\item{lang}{language to use. "en" = English; "es" = Spanish; "fr" = French.}
}
\value{
A plot of 50 percent and 90 percent confidence intervals of population projections if \code{spaghetti == FALSE} or a spaghetti plot with n individual projections if \code{spaghetti == n },  from \code{projections()}.
}
\description{
plots outputs from several projections that result from \code{projections()}, with multiple depletion levels.
}
\examples{
parms <- list(S0 = 0.944, S1plus = 0.99, 
K1plus = 9000, 
AgeMat = 18, nages = 20,  
z = 2.39, lambdaMax = 1.02)
InitDepl.vec <- c(0.1, 0.5, 0.9)
BycatchCV <- 0.2
nyears <- 100

high.list <- lapply(
X = InitDepl.vec,
function(x) {
  projections(
    NOut = 50,
    ConstantRateBycatch = list(Rate = 0.3, CV = BycatchCV),
    InitDepl = x,
    lh.params = parms,
    nyears = nyears,
    obs_CV = 0.2
  )
}
)
med.list <- lapply(
X = InitDepl.vec,
function(x) {
  projections(
    NOut = 50,
    ConstantRateBycatch = list(Rate = 0.02, CV = BycatchCV),
    InitDepl = x,
    lh.params = parms,
    nyears = nyears,
    obs_CV = 0.2
  )
}
)
low.list <- lapply(
X = InitDepl.vec,
function(x) {
  projections(
    NOut = 50,
    ConstantRateBycatch = list(Rate = 0.001, CV = BycatchCV),
    InitDepl = x,
    lh.params = parms,
    nyears = nyears,
    obs_CV = 0.2
  )
}
)
multiplot_proj(
high.d1 = high.list[[1]], # d1 is the lowest depletion
med.d1 = med.list[[1]],
low.d1 = low.list[[1]],
high.d2 = high.list[[2]],
med.d2 = med.list[[2]],
low.d2 = low.list[[2]],
high.d3 = high.list[[3]],
med.d3 = med.list[[3]],
low.d3 = low.list[[3]],
years.to.plot = nyears
)
}
