% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_plot_yield_curve.R
\name{plot_yield_curve}
\alias{plot_yield_curve}
\title{Yield curve for Shiny}
\usage{
plot_yield_curve(lh.params, z, MNPL_in, lang = "en")
}
\arguments{
\item{lh.params}{a list of life history parameters}

\item{z}{a vector of degrees of compensation that you want to plot. Default is the user-defined z, calculated from \code{MNPL_in}}

\item{MNPL_in}{the value of Maximum Net Productivity Level (MNPL) that results in}
}
\value{

}
\description{
Yield curve for Shiny
}
\examples{
plot_yield_curve(lh.params = list(S0 = 0.944, S1plus = 0.99, AgeMat = 17, 
nages = 19,  fmax = 0.29, z = 2.39, lambdaMax = 1.04, K1plus = 9000), 
z = 2.39, MNPL_in = 0.5, lang = "en")
}
