% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/23_prob_rebuilt_goal.R
\name{prob_rebuilt_goal}
\alias{prob_rebuilt_goal}
\title{Calculate the probability of reaching a rebuilding goal}
\usage{
prob_rebuilt_goal(traj, goal = 2000, rebuild.yr = 100)
}
\arguments{
\item{traj}{- a matrix of trajectories, with rows=nsims and cols=nyears}

\item{goal}{- what is the rebuilding goal?}

\item{rebuild.yr}{- the year by which you want pop to be recovered (calculate probability that pop will recover to MNPL by rebuild.yr)}
}
\value{
- the probability that the stock will be rebuilt to the goal population size by year \code{rebuilt.yr}
}
\description{
- a simple function that calculates the probability of recovery (in terms of 1+ numbers) to MNPL (or other population size) at \code{rebuilt.yr} years
}
\examples{
#' lh.params = list(S0 = 0.944, S1plus = 0.99, AgeMat = 17, 
PlusGroupAge = 19,  z = 2.39, lambdaMax = 1.04, K1plus = 9000)
Projection <- projections(NOut=100,ConstantBycatch = list(Catch = 200, CV = 01),InitDepl=0.8,lh.params=lh.params,nyears=40,obs_CV = 0.2) 
print(prob_rebuilt_goal(Projection$trajectories, goal = 0.5 * lh.params$K1plus, rebuild.yr = 40))

}
