% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/21_projections.R
\name{projections}
\alias{projections}
\title{Run simulations of a marine mammal population with bycatch mortality}
\usage{
projections(
  NOut,
  ConstantBycatch = list(Catch = NA, CV = NA),
  ConstantRateBycatch = list(Rate = NA, CV = NA),
  InitDepl,
  lh.params,
  nyears,
  obs_CV = 0
)
}
\arguments{
\item{NOut}{- number of simulations}

\item{ConstantBycatch}{Mean and CV of number of animals killed as bycatch per year}

\item{ConstantRateBycatch}{Mean and CV of bycatch rate}

\item{InitDepl}{initial depletion. If obs_CV>0, this is the mean depletion.}

\item{lh.params}{- life history parameters as a list. The list must include S0, S1plus, K1plus, AgeMat, nages, z, and lambdaMax}

\item{nyears}{number of years to do projections}

\item{obs_CV}{observation CV. Default to 1 for simple projections}
}
\value{
list of outputs from simulations: \code{params} contains parameter values for each trajectory; \code{trajectories} matrix contains simulation outputs; \code{fishing.rates} contain the bycatch rates for each year in each simulation; \code{InitDepl} returns the initial depletion for the projections; \code{ConstantBycatch} provides Catch (total individuals killed in bycatch events per year) and CV of Catch (if the user has specified bycatch as a constant number); \code{ConstantRateBycatch} contains Bycatch Rate (additional mortality from bycatch each year)  and CV of ByCatch rate. Other parameters are the same as in the \code{dynamics()} function.
}
\description{
- generates several projections, stochasticity is in the number of catches from year to year
}
\examples{
projections(NOut = 3, ConstantRateBycatch = list(Rate = 0.01, CV = 0.3),
InitDepl = 0.8,
lh.params = list(S0=0.944,S1plus=0.99,
K1plus=9000,AgeMat=18,nages=20,z= 2.39,lambdaMax=1.02),
nyears = 50, obs_CV = 0)

}
